Fix: Rename Number\Exact to Number
